style: #60 폴더구조 나누기
