+++ Spoony-iOS/Spoony-iOS/Source/Feature/Home/Search/Models/SearchResult.swift
@@ -2,7 +2,7 @@
 //  SearchResult.swift
 //  Spoony-iOS
 //
-//  Created by 이지훈 on 1/17/25.
+//  Created by 이지훈 on 1/16/25.
 //
 
 import Foundation
