+++ Spoony-iOS/Spoony-iOS/Source/Feature/Home/Search/SearchState.swift
@@ -1,0 +1,14 @@
+//
+//  SearchState.swift
+//  Spoony-iOS
+//
+//  Created by 이지훈 on 1/16/25.
+//
+
+import Foundation
+
+enum SearchState {
+    case empty
+    case typing
+    case searched
+}
